# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise AI Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List requests
      parameters:
        - name: module
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Request"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a new request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestResult"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/simulate:
    post:
      operationId: simulateRequest
      tags: [requests]
      summary: Simulate AI routing for a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateInput"
      responses:
        "200":
          description: Classification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Request:
      type: object
      required: [id, userId, title, details, module, risk, status, approvalRequired, createdAt, summary]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        details:
          type: string
        module:
          type: string
        risk:
          type: string
        status:
          type: string
        approvalRequired:
          type: boolean
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
        summary:
          type: string
    RequestInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
    RequestResult:
      type: object
      required: [ok, module, approval]
      properties:
        ok:
          type: boolean
        module:
          type: string
        approval:
          type: boolean
        requestId:
          type: ["integer", "null"]
    SimulateInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
    ClassificationResult:
      type: object
      required: [module, risk, approvalRequired, summary]
      properties:
        module:
          type: string
        risk:
          type: string
        approvalRequired:
          type: boolean
        summary:
          type: string
...