# lib/api-zod/src/generated/types/requestResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise AI Operations Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestResult {
  ok: boolean;
  module: string;
  approval: boolean;
  /** @nullable */
  requestId?: number | null;
}